MicroService_2/requirements.txt
Flask==2.3.2
SQLALchemy==2.0.20
psycopg2-binary==2.9.7
request==2.31.0
unittest-xml-reporting==3.0.4